requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
WTForms==3.2.1
python-dotenv==1.0.1
SQLAlchemy==2.0.38
pyodbc==5.2.0
pandas==2.2.3
numpy==2.2.4
scikit-learn==1.6.1
joblib==1.4.2
matplotlib==3.10.1
openpyxl==3.1.5
gunicorn==23.0.0
